add base reflection test code for checkerboard cube map
in a nutshell, we can choose to render the frame model map by doing the
6 pass draw to each of the cubemap's textures, OR we can render the
additional sphere and assess how that looks. This is solely test code to
verify a number of assumptions that were crucial for debugging issue #4.
One of these was that the cube map vector in the reflection shader was
correct; a simple way of verifying this is to use a very basic texture
with little distortion and simple patterns.