ensure that uniform_names are used when uniforms is empty
(and indentation). uniform_names is used to keep storage low, since
duniforms are unions that each take up 64 bytes + the storage needed for
the uniform name, in size. After we've stored the initial values in the
backing, we can simply refer to them using the name only - it's not the
render pass's job to care about what values each uniform is set to.