add shader generation routines
since lighting will be added soon (as shown by the macros defined at the
top of the source file), reusing code definitions is going to be
difficult without some kind of established methodology for this.

While a true preprocessor would be the most preferred, it will take time
to implement this.

Since we're generating the shader code at runtime, it's easiest to
simply include various code "chunks" on a per-component basis (as
indicated by the provided flag declarations).

Lighting will of course be one of these kinds of code chunks.