declarations
note that the GLSL_I* macros are for passing string literals
as arguments directly: the alternative - typing raw GLSL syntax
and using the preprocessor's string conversion operator after-the-fact -
is preferred. However, this isn't always possible when appending
external strings into GLSL expressions, given that that the preprocessor
can get confused by parenthesis which appear to be closing the macro
invocation, but are really meant to be a part of the GLSL expression.