initialize out_color, and rearrange its usage
GLSL doesn't zero out memory necessarily (this could be driver-dependent
and undefined as far as spec is concerned). out_color will be used for
assignments/blending operations that don't have anything to do with the
input frag_Color; we'll distribute the frag_Color over the final
calculation at the very end.

(note that, yes, floating point arithmetic is fundamentally not
commutative or associative, so variance in distribution can produce
different results - in this case, there isn't much point YET in dealing
with this, but it's worth mentioning).