uniform model transformation support in fragment shader
there are two ways we can deal with passing optional vertex data to the
fragment shader.

The first is to transform position and normal values in the vertex
shader. In this case, interpolation happens over already transformed
vertex data - the interpolation itself is just weighted and bound to the
actual triangle vertices themselves, so in this sense it's irrelevant
what space this happens in.

The second approach is to simply pass the vertices through the vertex
shader and perform the transformation directly in the fragment shader,
post interpolation. This is less efficient, since obviously a fragment
shader is executed on a per-fragment and not per-vertex basis. However,
this can be useful in certain situations, such as debugging.

There's a proof on the wiki page showing that both approaches are
equivalent.