fixes for bugfix/envmap-pointlights
1) FOV for projection matrix used in the environment map FBO needs to be
smaller. If it's too large, then the cube map reflection is distorted
such that the reflection of each face does not have matching edges with
its neighboring reflections.

2) The color buffer and depth buffers _must_ be cleared for the color
attachment and depth attachment respectively, to ensure that the
appropriate data is written on the draw call. The depth attachment is
initialized with depth values of 0, and the color attachment is
initialized with debug data. This explains why the pointlights weren't
actually being shown in the first place: all of the fragments were being
discarded, and what was shown was the initial texture that all cube maps
are provided with when created - this is the same texture as that of the
test skybox used, so it's easy enough to confuse the results.

3) Don't disable the first pass after it's been rendered; this was used
for testing and is no longer necessary.