bug fix for mouse-object movement logic
a simple point plane project is all that's necessary. There are still
issues, but the object can at least be moved. Note that s2w.x/s2w.y are
both multiplied with constants to improve the movement. For the X-axis,
10 appears to be a good value, but these coefficients are also dependent
on things like the distance from the object to the viewer and the actual
position of the object itself. This is primarily due to the limitations
of perspective projection.