mouse movement modifications
these changes apply somewhat to user object picking as well. A lambda is
introduced to group part of the mouse computations, since they'll be
used in both conditional statements.

Note that the g_cam_orient.dy calculation has an inverted sign from
before - this ensures that the vertical mouse movement does not produce
an inverted vertical camera orientation.

The addition of sdx/sdy are used to determine the sign of the mouse
movement; e.g., whether or not horizontal movement is to the right or
left, and whether or not vertical movement is upward or downward. This
will be useful in testing issues corresponding to movement of picked
objects.