extra state pipeline configuration
note that the spec for the  FRAMEBUFFER_SRGB extension implies that this
isn't necessary for framebuffer objects as long as the color attachment
isn't in SRGB format:

"If FRAMEBUFFER_SRGB is enabled and the value of
    FRAMEBUFFER_ATTACHMENT_COLOR_ENCODING for the framebuffer attachment
    corresponding to the destination buffer is SRGB (see section 6.1.3),
    the R, G, and B values after blending are converted into the
non-linear
    sRGB color space"

DESPITE this, the more explicit we are in how we handle state
configuration/queries, the better. It reduces the chances of bugs
occurring (both on our end and in the OpenGL driver).

The mouse picking buffer doesn't encode its data in SRGB.

That quote can be found toward the end of "Additions to Chapter 4 of the
2.0 Specification (Per-Fragment Operations
and the Frame Buffer)", here:
https://www.khronos.org/registry/OpenGL/extensions/ARB/ARB_framebuffer_sRGB.txt