toggle quad: some fragment shader uniforms that are useful for debugging
I was using this to verify that the XY coordinates were being
interpreted exactly as I thought they would be when I clicked
the screen. It turns out they weren't, and this has lead to the fix for
the mouse picking issue of the last few days. The picking is now
accurate.