add a plane struct and some function overloads
also a convenience yz plane definition - others of course can be added
in a similar fashion.

(this is for issues pertenant too the camera orientation: in particular,
this helps with locking the orientation when both the Y and X axes have
been rotated about - NOTE: technically, this still needs work, but the
plane definitions are useful regardless).