MAIN FIX: camera orientation (see detailed explanation in commit message)
before this, the y mouse coordinate position was being read as-is, and
operations were being performed on it from there.

Then, the height of the framebuffer minus that value would be stored in
the prev_ypos value (this was to ensure that the mouse cursor position
was being properly mapped to the screen).

The end result of this is an always negative dy rotation value (which is
the delta accumulation angle that rotates about the x-axis for the
camera's orientation transform).

For a more mathematical explanation,

let y_i be the current yposition received in the function,
    dy_i be our delta y (g_cam_orient.dy),
    U be our previously computed y position (g_cam_orient.prev_ypos),
    and H our framebuffer height

Upon entry of the function, we have:

dy_i = y_i - U
     = y_i - (H - y_{i - 1})
     = (y_i + y_{i - 1}) - H
     < 0 (always)

Updating U,

U = H - y_i
-------------
Now, instead of this, define an additional py_i, along with the previous
definitions:

py_i = H - y_i
dy_i = py_i - U
     = py_i - py_{i - 1}
     = (H - y_i) - (H - y_{i - 1})
     = (H - H) + (y_{i - 1} - y_i)
     = y_{i - 1} - y_i

If we removed H and instead simply let py_i = y_i, and computed the
above, we'd have dy_i = y_i - y_{i - 1}.

Now, while for _camera orientation_ the inclusion of the H in the
computation makes little difference, the y position itself still must
have this for the actual mouse picking to work properly, since GLFW's y
axis increases downward, and OpenGL's viewport increases upward.