render pass containers fix
Currently, the way the render passes are ordered is important: one pass
facilitates the clearing of the frame and depth buffers (primarily -
there may be exceptions, e.g. a subsequent pass might need the depth
buffer cleraed to guarantee it's rendered in front of everythign else).

The map is used instead of a vector here since this will be changed in
the future, and that will ease modifications to any loops which iterate
over these passes.