first commit toward nuklear implementation
this is a behemoth of a commit: as it stands, time is limited, so I'm
going to hold off on breaking this one up.

In a nutshell:

- all of the GLFW/OpenGL initialization code has been moved over to
device_context.hpp

- device_context.hpp contains support for fence synchronization

- the overall module init process has been streamlined, so that recovery
paths are easier to mentally parse

- there is a bugfix in the cursor_position_callback() that includes the
addition of missing parenthesis for testdy; these had been there before
originally (as soon as yesterday), but somehow were erased. this was
really screwing up mouse movement. see line 1156 of main.cpp for
context.

base/gui.cpp and base/gui.hpp will be the primary driver for the GUI.
Ideally, most of the nuklear state will be stored there.