new gapi handle type system definitions/changes
in addition to a lot of cleanup.

handle_int_t is now a 64 bit signed integer instead of an unsigned 32
bit integer. This is more flexible, since with OpenGL, null values can
be either 0 or -1 depending on the type of call being used.

There is also a new type handle generator, that makes reading the code
significantly easier and reduces some of the macro obfuscation.

Also some operator == overload generator macros, since more null types
are defined. For an example, see the included change for gapi.cpp in
this commit.