use set_null and default initializers
this is important, since using values like "0" is different from the
actual null value that's used in the handle type. The only constants we
should be using in this context are constants that are involved in value
comparisons/checks against OpenGL function calls.

An example is if glCreateShader() fails, then 0 should be returned.