add DEVICE_HANDLE_OPS_MT macro
useful for dealing with handles that can be bound to different kinds of
targets (especially texture and buffer objects - framebuffer objects
technically apply as well here, but less so).

Note that the newly added enum_type class in util.hpp is needed for this
new macro: it iterates over enum values that correspond to a given
target type, uses each value as a key into the map that's generated by
the macro.

The enum_type class acts as a pseudo-reflection utility for this very
purpose (this can also only be used if enum class types have
enum_type_first and enum_type_last member values defined).