this ensures that vulkan::renderer::~renderer() gets called
since renderloop_triangle inherits from renderloop. There is no explicit
destructor defined, but the implicit destructor will be inserted and
actually placed in the renderloop vtable with this modification. This in
turn will cause the compiler to insert a a call to
vulkan::renderer::~renderer() inside of renderloop_triangle's
destructor.