vulkan utility function for selecting appropriate device memory source
in a nutshell, pretty much any resource (i.e., a VkBuffer or VkImage)
has a series of requirements that are dependent on the desired size of
the resource itself (in terms of memory).

these requirements include a valid device memory source that
is compatible with the requested size of the source.

this function is designed to find the device memory source which holds
the properties provided by req_properties, choosing from memory sources
that are available on the device (via memory_type_bits_req).

Note that the memory_type_bits_req may differ, depending on the resource
in question (in terms of size, format, etc.), so it's important to
always create a dummy resource _first_, and then use the Vulkan API to
query the various parameters the driver expects in terms of its storage.