first set of code for 2d texture creation
Currently this allocates image memory, and then binds that memory to an
image resource, while also creating a corresponding view, and ensuring
that the requested sizes meet the memory requirements returned from a
query to the driver.

What is left to do is figure out how to actually upload the texture data
to the storage backed by the image device memory.