validate our image create info
Vulkan's validation layer will report an error if the image info's
format, image type, tiling method, usage, and flag values are
incompatible.

We perform this same check here manually. The reason for this is that
(a) we might later have to come up with some kind of format mapping that
can be cycled through until an appropriate combination is found, and
(b) catching issues here in the future will help prevent
the spread of further issues down the road.

(further calls to api_ok() will return false if there is a
problem)