allocate descriptor pool & pre/post conditions
we have another event that isn't yet committed: setup_texture_data().
This is placed in between setup_descriptor_pool() and
setup_graphics_pipeline(). It's left out of this
commit, because there are other modifications that should be noted
before this it's provided.

setup_graphics_pipeline() is modified to expect ok_texture_data() to
return true _before_ it begins its own execution. This won't happen
until the corresponding code is provided.