add support for flying camera
---
device_context.hpp:
---
success is only set to true in the event that we're using opengl without
this modification; until now, it wasn't really relevant. At this point,
we need it to register input callback handlers.
---
main.cpp:
---
take the initialization of g_m.view out of the "only opengl" case in the
modules initialization routine, and initialize it regardless of the API
being used. We naturally delete the view unconditionally as well.

the view's projection matrix must be explicitly initialized, so we call
g_m.view->reset_proj() for this. within the render_loop_triangle::init()
routine.

we also add appropriate update call and pass updated results (view and
projection transforms) to vulkan through the uniform block.
---
vk_uniform_buffer.hpp:
---
the device memory has to be updated regularly, so in order for us to do
that (while ensuring that the user who wishes to update said memory only
  need to pass the index returned from the "add()" method), we add
another user storage function which stores the actual user pointer
pointing
to said block data.

We also refactor the initial device memory map code into its own
method, since we're using it in two locations (the first is during
uniform block initialization, the second is when we update the memory).
---
vulkan.hpp:
---
we add a block structure (transform_data) which directly maps to the
uniform block that we've defined inside the vertex shader. we also add a
few set_* methods for the matrices in the actual transform_data block
(that's also used by m_vertex_uniform_block). we also add a
corresponding update_block() call in the render() method, so that our
view state is always taken care of after it's changed.

we also change the y-axis on the triangle vertices, since we're handling
the differences between vulkan/opengl on that front in the vertex
shader.
---
tri_ubo.vert.glsl:
---
as we can see, the uniform buffer (also represented client side by
transform_data) is defined with appropriate matrices.

we can see that gl_Position is written to 3 times here. The first time
is the standard vertex -> view -> projection transformation process.

The second time is due to the fact that vulkan's y-axis increases
downward (opengl's increases upward).

The third time is to guarantee a normalization of depth between [-1, 1]
to [0, 1]: vulkan's NDC space z-axis is in [0, 1] (OpenGL's is [-1, 1]).