two triangles and a working depth buffer
The two triangles is used as a test for the depth buffer. Changing the
colors makes it clear which triangle is "winning" the depth test and
which isn't, assuming there is a z-conflict in the first place.

-------
vulkan.cpp: disable back face culling
-------
vulkan.hpp: defer differences in triangles to shaders/vertex data.

In a nutshell,
- added a second set of vertices
- draw commands use only one render pass
- instance indices are now used to communicate to shaders which
triangle is being drawn (in vkCmdDraw())
-------
main.cpp: set the camera so the triangles are displayed in front of the
user
-------
tri_ubo.vert.glsl: add color pass through and determine which triangle
is being drawn via gl_InstanceIndex.
-------
tri_ubo.frag.glsl: add color pass through from vertex shader
-------